backend/requirements.txt
# FastAPI & ASGI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# MongoDB
motor>=3.3.0
pymongo>=4.6.0

# Auth & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Config & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Email validation
email-validator>=2.1.0
